ADD useShoppingCart custom hook
