+++ src/component-patterns/hooks/useProduct.ts
@@ -1,4 +1,4 @@
-import { useEffect, useState } from 'react';
+import { useEffect, useRef, useState } from 'react';
 import { OnChangeArgs, Product } from '../interfaces/interfaces';
 
 interface useProductArgs {
@@ -14,15 +14,21 @@ export const useProduct = ({
 }: useProductArgs) => {
   const [counter, setCounter] = useState(value);
 
+  const isControlled = useRef(!!onChange);
+
   const increaseBy = (value: number) => {
+    if (isControlled.current) {
+      return onChange!({ count: value, product });
+    }
+
     const newValue = Math.max(counter + value, 0);
     setCounter(newValue);
     onChange && onChange({ count: newValue, product });
   };
 
   useEffect(() => {
-    setCounter(value)
-  }, [ value ]);
+    setCounter(value);
+  }, [value]);
 
   return {
     counter,
