+++ src/component-patterns/interfaces/interfaces.ts
@@ -26,3 +26,7 @@ export interface OnChangeArgs {
   product: Product;
   count: number;
 } 
+
+export interface ProductInCart extends Product {
+  count: number;
+}
